Add Node MultiGeometry with highlighting paths
Changed Node Placemark from just a Point to a MultiGeometry with Point and LineString.  
Moved Link/Paths loop inside Node Placemark, so we can use it to also write the LineString for the Node Placemark.  
Added code to generate the new LineString.  Not sure if I'm getting the node2 info correctly (copied from onode in Path code).