Improve path type code & styles
Refactored section that detects path type, to simplify and only generate new path_type variable, which is then used to write Placemark name and translated to path's styleUrl. This involved moving the Push that writes the Path Placemark KML out of the switch, to the parent IF. 
Update styles to remove labels from Path multigeometries.